Fix copyCompatibility bug in copy method for checkbox column data
When the option copyCompatibility: true, there is currently a bug where it copies the innerHTML rather than value of TRUE/FALSE.

The fix is to check the column type, and modify the treatment of the text that is copied into the clipboard.

In the case of checkbox, we should NOT do the previous default behavior of copying the innerHTML, instead, we should copy the value of TRUE or FALSE. However, if the default innerHTML has been override pragmatically by the user, then we should revert to default treatment of copying the innerHTML, as the value can no longer be relied on being the intended copy source.

This patch maintains backwards compatibility, while adding a new fork in the control flow to better handle different column data-types, such as: image, color bar, etc. The intent is to improve ease of copy/pasting data into ASCII that can be understood by most spreadsheet apps when the data is pasted into them. At the moment, this patch is only proposing a solution to handling the checkbox. Future patches can be easily coded to handle other cases noted above.